refactor: 긴급대응 관련 폴더구조 변경 design: 반투명 버튼 및 반투명 border에 css 적용
