+++ src/pages/Emergency/EmergencyMain/OrganCard.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 type Props = {
   organData: {
     id: number;
