+++ src/pages/Emergency/EmergencyMain/QuickReportCard.tsx
@@ -1,4 +1,4 @@
-import Button from "../../components/Button/Button";
+import Button from "../../../components/Button/Button";
 
 type Props = {
   title: string;
